% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_sim123.R
\docType{data}
\name{example_sim123}
\alias{example_sim123}
\title{Example sim123 Dataset}
\format{
A list containing 3 data frames:
\describe{
  \item{smat}{A 1089 x 3 numeric matrix of the true source signals. Each column is an 33 x 33 image.}
  \item{mmat}{A 3 x 50 numeric mixing matrix of the true time series. Each row is a time series of corresponding column in \code{smat}.}
  \item{xmat}{A 1089 x 50 numeric matrix of the simulated data. Each column is the simulated mixed signal at a time point.}
}
}
\usage{
example_sim123
}
\description{
A simple dataset for demonstration purposes.
}
\examples{
data(example_sim123)
str(example_sim123)
}
\keyword{datasets}
