% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseICA-package.R
\name{gen.inits}
\alias{gen.inits}
\title{Function for generating random starting points}
\usage{
gen.inits(p, d, runs, orth.method = c("svd", "givens"))
}
\arguments{
\item{p}{Number of rows}

\item{d}{Number of columns}

\item{runs}{the number of random starts}

\item{orth.method}{orthodox method}
}
\value{
a list of initialization of matrices.
}
\description{
Function for generating random starting points
}
\examples{
gen.inits(2,3,3,"svd")
}
