% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{givens.rotation}
\alias{givens.rotation}
\title{For a given angle theta, returns a d x d Givens rotation matrix}
\usage{
givens.rotation(theta = 0, d = 2, which = c(1, 2))
}
\arguments{
\item{theta}{The value of theta.}

\item{d}{The value of d.}

\item{which}{The value of which.}
}
\description{
For a given angle theta, returns a d x d Givens rotation matrix
}
