% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseICA-package.R
\name{matchICA}
\alias{matchICA}
\title{Match ICA results based on L2 distances and Hungarian}
\usage{
matchICA(S, template, M = NULL)
}
\arguments{
\item{S}{loading variable matrix}

\item{template}{template for match}

\item{M}{subject score matrix}
}
\value{
the match result
}
\description{
Match ICA results based on L2 distances and Hungarian
}
