% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{signchange}
\alias{signchange}
\title{Change the sign of S and M matrices to positive skewness.}
\usage{
signchange(S, M = NULL)
}
\arguments{
\item{S}{The S matrix with dimension P x Q.}

\item{M}{The M matrix with dimension Q x T.}
}
\value{
A list of S and M matrices with positive skewness.
}
\description{
Change the sign of S and M matrices to positive skewness.
}
