% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SICAmain.R
\name{soft_thresh_R}
\alias{soft_thresh_R}
\title{Soft-threshold function}
\usage{
soft_thresh_R(x, nu = 1, lambda = sqrt(2)/2)
}
\arguments{
\item{x}{The input scalar.}

\item{nu}{The tuning parameter.}

\item{lambda}{The lambda parameter of the Laplace density.}
}
\description{
Soft-threshold function
}
