% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseICA-package.R
\name{sparseICA}
\alias{sparseICA}
\title{Sparse ICA function}
\usage{
sparseICA(
  xData,
  n.comp,
  nu = 1,
  U.list = NULL,
  whiten = c("eigenvec", "sqrtprec", "lngca", "none"),
  orth.method = c("svd", "givens"),
  method = c("C", "R"),
  restarts = 40,
  lambda = sqrt(2)/2,
  irlba = FALSE,
  eps = 1e-06,
  maxit = 500,
  verbose = FALSE,
  converge_plot = FALSE,
  col.stand = TRUE,
  row.stand = FALSE,
  iter.stand = 0
)
}
\arguments{
\item{xData}{Input data matrix with dimension P x T. P is the number of features. t is the number of samples.}

\item{n.comp}{The number of components.}

\item{nu}{the tuning parameter controlling the accuracy and sparsity of the results. Should be selected by the BIC-like criterion "BIC_sparseICA_R()" or expert knowledge. The default is 1.}

\item{U.list}{The initialization of U matrix. Default is "NULL".}

\item{whiten}{The method for whitening input xData. Could take "eigenvec", "sqrtprec", "lngca", and "none". The default is "eigenvec".}

\item{orth.method}{The method used for generating initial values of U matrix. The default is "svd".}

\item{method}{If method == "C" (default) then C code is used to perform most of the computations, which makes the algorithm run faster. If method == "R" then computations are done exclusively in R.}

\item{restarts}{The number of initial points. Default is 40.}

\item{lambda}{The scale parameter in Laplace density. The default is sqrt(2)/2 to make the default situation with unit variance.}

\item{irlba}{Whether use the irlba method to perform fast truncated singular value decomposition in whitening step. The default is FALSE.}

\item{eps}{The convergence threshold. The default is 1e-6.}

\item{maxit}{The maximum number of iterations in the Sparse ICA method with Laplace density. The default number is 500.}

\item{verbose}{Whether print the information about convergence. The default is FALSE.}

\item{converge_plot}{Whether make a convergence plot for Sparse ICA method. The default is FALSE.}

\item{col.stand}{Whether standardize the data matrix column-wise. Default if TRUE.}

\item{row.stand}{Whether standardize the data matrix row-wise. Default if FALSE.}

\item{iter.stand}{The number of standardization. Default is 0.}
}
\value{
Function outputs a list including the following:
\describe{
      \item{\code{loglik_restarts}}{The value of minimal log-likelihood among random initializations.}
      \item{\code{estS}}{The estimated sparse independent components matrix with dimension P x Q.}
      \item{\code{estU}}{The estimated U matrix with dimension Q x Q.}
      \item{\code{converge}}{The convergence of the U matrix.}
      \item{\code{distribution}}{The density used in ICA method.}
      \item{\code{whitener}}{The whitener matrix used to perform data whitening.}
      \item{\code{estM}}{Estimated mixing matrix with dimension Q x T.}}
}
\description{
The function to perform Sparse ICA, which is implemented in both pure R and RCpp.
}
\examples{
\donttest{
#get simulated data
data(example_sim123)

my_sparseICA = sparseICA(xData = xmat, n.comp = 3, nu = 1.2, U.list=NULL,
    whiten = "eigenvec", orth.method = "svd", method="C",restarts = 40,
    lambda = sqrt(2)/2, eps = 1e-6, maxit = 500, verbose=TRUE)

a=matchICA(my_sparseICA$estS,smat)

par(mfrow=c(1,3))
image(matrix(a[,1],33,33))
image(matrix(a[,2],33,33))
image(matrix(a[,3],33,33))
par(mfrow=c(1,1))
}
}
