% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{theta2W}
\alias{theta2W}
\title{Convert angle vector into orthodox matrix}
\usage{
theta2W(theta)
}
\arguments{
\item{theta}{A vector of angles theta.}
}
\value{
An orthodox matrix.
}
\description{
Convert angle vector into orthodox matrix
}
