% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SICAmain.R
\name{whitener}
\alias{whitener}
\title{The function for perform whitening.}
\usage{
whitener(X, n.comp = ncol(X), center.row = FALSE, use_irlba = TRUE)
}
\arguments{
\item{X}{The data matrix with dimension P x T.}

\item{n.comp}{The number of components.}

\item{center.row}{Whether to center the row of data. Default is FALSE.}

\item{use_irlba}{Whether to use the irlba method to perform fast truncated singular value decomposition in whitening step, helpful for memorying intermediate dataset. Default is TRUE.}
}
\value{
A list including the whitener matrix, the whitened data matrix, and the mean of the input data.
}
\description{
The function for perform whitening.
}
