% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseICA-package.R
\name{BIC_sparseICA}
\alias{BIC_sparseICA}
\title{BIC-like selection criterion function}
\usage{
BIC_sparseICA(
  xData,
  n.comp,
  nu_list = seq(0.1, 4, 0.1),
  U.list = NULL,
  whiten = c("eigenvec", "sqrtprec", "lngca", "none"),
  orth.method = c("svd", "givens"),
  method = c("C", "R"),
  lambda = sqrt(2)/2,
  irlba = FALSE,
  eps = 1e-06,
  maxit = 500,
  verbose = FALSE,
  col.stand = TRUE,
  row.stand = FALSE,
  iter.stand = 0,
  BIC_plot = FALSE
)
}
\arguments{
\item{xData}{Input data matrix with dimension P x T. P is the number of features. t is the number of samples.}

\item{n.comp}{The number of components.}

\item{nu_list}{the list of candidate tuning parameter. Default is seq(0.1,4,0.1).}

\item{U.list}{The initialization of U matrix. Default is "NULL".}

\item{whiten}{The method for whitening input xData. Could take "eigenvec", "sqrtprec", "lngca", and "none". The default is "eigenvec".}

\item{orth.method}{The method used for generating initial values of U matrix. The default is "svd".}

\item{method}{If method == "C" (default) then C code is used for Sparse ICA. If method == "R" then computations are done exclusively in R.}

\item{lambda}{The scale parameter in Laplace density. The default is sqrt(2)/2 to make the default situation with unit variance.}

\item{irlba}{Whether use the irlba method to perform fast truncated singular value decomposition in whitening step. The default is FALSE.}

\item{eps}{The convergence threshold. The default is 1e-6.}

\item{maxit}{The maximum number of iterations in the Sparse ICA method with Laplace density. The default number is 500.}

\item{verbose}{Whether print the information about convergence. The default is FALSE.}

\item{col.stand}{Whether standardize the data matrix column-wise. Default if TRUE.}

\item{row.stand}{Whether standardize the data matrix row-wise. Default if FALSE.}

\item{iter.stand}{The number of standardization. Default is 0.}

\item{BIC_plot}{Whether make a convergence plot for BIC selection. The default is FALSE.}
}
\value{
Function outputs a list including the following:
\describe{
      \item{\code{BIC}}{The list of BIC values corresponding to each candidate in the nu_list.}
      \item{\code{nu_list}}{The list of candidate tuning parameter nu.}
      \item{\code{best_nu}}{The best nu selected by BIC-like criterion.}}
}
\description{
Use BIC-like criterion to select tuning parameter nu in Sparse ICA.
}
\examples{
\donttest{
#get simulated data
data(example_sim123)

select_sparseICA = BIC_sparseICA(xData = xmat, n.comp = 3,U.list=NULL,
    whiten = "eigenvec", orth.method="svd", method="C", lambda = sqrt(2)/2,
    eps = 1e-6,maxit = 500, BIC_plot = TRUE,verbose=FALSE,
    nu_list = seq(0.1,4,0.1))

my_nu = select_sparseICA$best_nu
}
}
