% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SICAmain.R
\name{gen.inits}
\alias{gen.inits}
\title{Function for generating random starting points}
\usage{
gen.inits(p, d, runs, orth.method = c("svd", "givens"))
}
\arguments{
\item{p}{The number of rows.}

\item{d}{The number of columns.}

\item{runs}{The number of random starts.}

\item{orth.method}{The method used for generating initial values of U matrix. The default is "svd".}
}
\value{
A list of random initialization of matrices.
}
\description{
Function for generating random starting points
}
