% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseICA-package.R
\name{signchange}
\alias{signchange}
\title{Sign change for S matrix to image}
\usage{
signchange(S, M = NULL)
}
\arguments{
\item{S}{S matrix, P x Q.}

\item{M}{M matrix, Q x T.}
}
\value{
a list of positive S and positive M.
}
\description{
Sign change for S matrix to image
}
