% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseICA-package.R
\name{soft_thresh_R}
\alias{soft_thresh_R}
\title{Soft-thresholding to update V}
\usage{
soft_thresh_R(x, nu = 1, lambda = sqrt(2)/2)
}
\arguments{
\item{x}{input scalar}

\item{nu}{the tuning parameter}

\item{lambda}{the lambda parameter of the Laplace density}
}
\description{
Soft-thresholding to update V
}
