% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseICA-package.R
\name{whitener}
\alias{whitener}
\title{Whitening Function}
\usage{
whitener(X, n.comp = ncol(X), center.row = FALSE, irlba = FALSE)
}
\arguments{
\item{X}{The data matrix P x T}

\item{n.comp}{The number of components}

\item{center.row}{Whether center the row of data}

\item{irlba}{Whether use the irlba package}
}
\value{
a whitener matrix
}
\description{
Whitening Function
}
